Fixes for spaces in paths/PATHEXT for windows.
In `latex.ts` and `pdf.ts`, I've wrapped the paths for the .tex/.pdf/.svg files in double quotes so that (as is not uncommon in Windows) there won't be errors if these paths have spaces in them. Additionally, I added the `shell = true` argument to `spawn` because otherwise it does not respect the `PATHEXT` environment variable on Windows which can lead it to throw ENOENT errors for commands that would work on the command line.